requirements.txt
pylint
pytest
pytest-cov
torch
torchvision
matplotlib
seaborn
tqdm
scikit-learn
gradio
Pillow
kagglehub
python-dotenv